cloud_functions/customer_check/requirements.txt
Flask==2.2.2
functions-framework==3.2.0
google-cloud-datastore==2.8.1